+++ packages/css-parser-algorithms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes to CSS Parser Algorithms
 
+### 4.0.0
+
+_January 14, 2026_
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+- Updated [`@csstools/css-tokenizer`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer) to [`4.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/css-tokenizer/CHANGELOG.md#400) (major)
+
 ### 3.0.5
 
 _May 27, 2025_